git-svn-id: svn://192.168.1.10/public/CoffeeMud@1301 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Items/Armor/GlowingMageArmor.java
@@ -28,18 +28,14 @@ public class GlowingMageArmor extends StdArmor
 	
 	public boolean savable(){return false;}
 	
-	public void affectEnvStats(Environmental affected, EnvStats affectableStats)
-	{
-		affectEnvStats(affected,affectableStats);
-		if(amWearingAt(Item.INVENTORY))
-		   destroyThis();
-	}
-	
 	public boolean okAffect(Affect affect)
 	{
 		if(!super.okAffect(affect))
 			return false;
 
+		if((amWearingAt(Item.INVENTORY))&&(!amDestroyed()))
+		   destroyThis();
+		
 		MOB mob=affect.source();
 		if(!affect.amITarget(this))
 			return true;
